Extract shared ListController snapshot tests
