+++ EssentialFeediOSTests/Shared UI/snapshots/EMPTY_LIST_dark.png
